requirements.txt
numpy>=1.24
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pandas>=2.2.0
pyarrow>=16.0.0
